test: restore filterPlaces unit test separation
Agent-Logs-Url: https://github.com/STH-1-Class-One-Group/JamIssue/sessions/92868498-f5de-42cc-804c-a35967297b97
